26-allow-passwordless-login (#27)
* Removing basicAuth requirement

* Adding additional disclaimer

With the intended use for Helios to always be behind an authentication proxy, we will disable it as it always gets in the way. In the future, we may find another way to use it, but for now we can disable and revisit if there is enough demand in the future.

---------

Co-authored-by: app/github-actions <app/github-actions>
Co-authored-by: aldmbmtl <alex.e.hatfield@gmail.com>